+++ reference_apps/real_estate_mls/src/intents/design_editor/index.tsx
@@ -7,7 +7,6 @@ import { App } from "./app";
 
 async function render() {
   const root = createRoot(document.getElementById("root") as Element);
-
   root.render(
     <AppI18nProvider>
       <AppUiProvider>