+++ templates/data_connector/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./app_error";
-export * from "./footer";
-export * from "./header";