+++ templates/data_connector/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from "./data_table";
-export * from "./data_params";