+++ templates/gen_ai/src/context/index.ts
@@ -1,2 +1,0 @@
-export * from "./app_context";
-export * from "./use_app_context";